+++ PhotoGallery.Infrastructure/ImageProcessing/Templates/DownloadImageTransformTemplate.cs
@@ -1,0 +1,14 @@
+﻿using PhotoGallery.Application.Abstractions;
+using PhotoGallery.Application.DTOs.PhotoGallery.Application.DTOs;
+
+namespace PhotoGallery.Infrastructure.ImageProcessing.Templates
+{
+    public sealed class DownloadImageTransformTemplate : ImageTransformTemplate
+    {
+        public DownloadImageTransformTemplate(IImageProcessorFactory processorFactory) : base(processorFactory)
+        {
+        }
+
+        protected override bool ShouldCreateThumbnail(ImageProcessingOptionsDto options) => false;
+    }
+}